chore: move Interfaces folder
