+++ app/src/main/res/layout/item_saved_search_word.xml
@@ -19,8 +19,8 @@
         android:id="@+id/saved_search_word_text_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="10dp"
-        android:paddingEnd="20dp"
+        android:layout_marginStart="5dp"
+        android:paddingEnd="10dp"
         android:textSize="18sp"
         android:textStyle="bold"
         app:layout_constraintBottom_toBottomOf="parent"
